+++ controllers/mysql_controller.go
@@ -66,11 +66,6 @@ func (a *mySQLAdapter) getUserConfig() any {
 }
 
 func (a *mySQLAdapter) newSecret(s *aiven.Service) (*corev1.Secret, error) {
-	name := a.Spec.ConnInfoSecretTarget.Name
-	if name == "" {
-		name = a.Name
-	}
-
 	stringData := map[string]string{
 		"MYSQL_HOST":        s.URIParams["host"],
 		"MYSQL_PORT":        s.URIParams["port"],
@@ -82,17 +77,7 @@ func (a *mySQLAdapter) newSecret(s *aiven.Service) (*corev1.Secret, error) {
 		"MYSQL_REPLICA_URI": s.ConnectionInfo.MySQLReplicaURI,
 	}
 
-	// Removes empties
-	for k, v := range stringData {
-		if v == "" {
-			delete(stringData, k)
-		}
-	}
-
-	return &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{Name: name, Namespace: a.Namespace},
-		StringData: stringData,
-	}, nil
+	return newSecret(a, a.Spec.ConnInfoSecretTarget, stringData), nil
 }
 
 func (a *mySQLAdapter) getServiceType() string {