+++ tests/clickhouse_test.go
@@ -1,14 +1,11 @@
 package tests
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/aiven/aiven-operator/api/v1alpha1"
 	clickhouseuserconfig "github.com/aiven/aiven-operator/api/v1alpha1/userconfig/service/clickhouse"
@@ -93,9 +90,8 @@ func TestClickhouse(t *testing.T) {
 	assert.Equal(t, ipFilterAvn, ch.Spec.UserConfig.IpFilter)
 
 	// Secrets test
-	ctx := context.Background()
-	secret := new(corev1.Secret)
-	require.NoError(t, k8sClient.Get(ctx, types.NamespacedName{Name: name, Namespace: "default"}, secret))
+	secret, err := s.GetSecret(ch.GetName())
+	require.NoError(t, err)
 	assert.NotEmpty(t, secret.Data["HOST"])
 	assert.NotEmpty(t, secret.Data["PORT"])
 	assert.NotEmpty(t, secret.Data["USER"])