+++ tests/clickhouseuser_test.go
@@ -1,14 +1,11 @@
 package tests
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/aiven/aiven-operator/api/v1alpha1"
 )
@@ -38,6 +35,13 @@ spec:
   authSecretRef:
     name: aiven-token
     key: token
+
+  connInfoSecretTarget:
+    name: my-ch-user-secret
+    annotations:
+      foo: bar
+    labels:
+      baz: egg
 
   project: %[1]s
   serviceName: %[2]s
@@ -84,13 +88,14 @@ func TestClickhouseUser(t *testing.T) {
 	assert.Equal(t, userName, user.GetName())
 	assert.Equal(t, userAvn.Name, user.GetName())
 
-	ctx := context.Background()
-	secret := new(corev1.Secret)
-	require.NoError(t, k8sClient.Get(ctx, types.NamespacedName{Name: userName, Namespace: "default"}, secret))
+	secret, err := s.GetSecret(user.Spec.ConnInfoSecretTarget.Name)
+	require.NoError(t, err)
 	assert.NotEmpty(t, secret.Data["HOST"])
 	assert.NotEmpty(t, secret.Data["PORT"])
 	assert.NotEmpty(t, secret.Data["PASSWORD"])
 	assert.NotEmpty(t, secret.Data["USERNAME"])
+	assert.Equal(t, map[string]string{"foo": "bar"}, secret.Annotations)
+	assert.Equal(t, map[string]string{"baz": "egg"}, secret.Labels)
 
 	// We need to validate deletion,
 	// because we can get false positive here: