+++ tests/grafana_test.go
@@ -1,14 +1,11 @@
 package tests
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/aiven/aiven-operator/api/v1alpha1"
 	grafanauserconfig "github.com/aiven/aiven-operator/api/v1alpha1/userconfig/service/grafana"
@@ -101,9 +98,8 @@ func TestGrafana(t *testing.T) {
 	assert.Equal(t, ipFilterAvn, grafana.Spec.UserConfig.IpFilter)
 
 	// Secrets test
-	ctx := context.Background()
-	secret := new(corev1.Secret)
-	require.NoError(t, k8sClient.Get(ctx, types.NamespacedName{Name: name, Namespace: "default"}, secret))
+	secret, err := s.GetSecret(grafana.GetName())
+	require.NoError(t, err)
 	assert.NotEmpty(t, secret.Data["GRAFANA_HOST"])
 	assert.NotEmpty(t, secret.Data["GRAFANA_PORT"])
 	assert.NotEmpty(t, secret.Data["GRAFANA_USER"])