+++ tests/opensearch_test.go
@@ -1,14 +1,11 @@
 package tests
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/aiven/aiven-operator/api/v1alpha1"
 	opensearchuserconfig "github.com/aiven/aiven-operator/api/v1alpha1/userconfig/service/opensearch"
@@ -24,6 +21,13 @@ spec:
   authSecretRef:
     name: aiven-token
     key: token
+
+  connInfoSecretTarget:
+    name: my-os-secret
+    annotations:
+      foo: bar
+    labels:
+      baz: egg
 
   project: %[1]s
   cloudName: google-europe-west1
@@ -96,11 +100,12 @@ func TestOpenSearch(t *testing.T) {
 	assert.Equal(t, ipFilterAvn, os.Spec.UserConfig.IpFilter)
 
 	// Secrets test
-	ctx := context.Background()
-	secret := new(corev1.Secret)
-	require.NoError(t, k8sClient.Get(ctx, types.NamespacedName{Name: name, Namespace: "default"}, secret))
+	secret, err := s.GetSecret(os.Spec.ConnInfoSecretTarget.Name)
+	require.NoError(t, err)
 	assert.NotEmpty(t, secret.Data["HOST"])
 	assert.NotEmpty(t, secret.Data["PORT"])
 	assert.NotEmpty(t, secret.Data["USER"])
 	assert.NotEmpty(t, secret.Data["PASSWORD"])
+	assert.Equal(t, map[string]string{"foo": "bar"}, secret.Annotations)
+	assert.Equal(t, map[string]string{"baz": "egg"}, secret.Labels)
 }